+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add `\Lullabot\Mpx\AuthenticatedClient::setTokenDuration` to set a limit for
+  token lifetimes.
+
 ## [0.2.0] - 2018-05-18
 
 ### Added